move functions to coordinator_service.rs
